Splitting gateway.js test scripts out from index.test.js. #5
